+++ src/lib/TextBlock/TextBlock.stories.svelte
@@ -11,8 +11,23 @@
         default: "body",
         options: ["body", "full"],
         control: "select"
+      },
+      variant: {
+        default: "primary",
+        options: ["primary", "reverse"],
+        control: "select"
       }
+
     },
+    parameters: {
+      backgrounds: {
+        default: "light",
+        values: [
+          { name: "light", value: "#ffffff" },
+          { name: "dark", value: "#0A4C6A" }
+        ]
+      }
+    }
       
 	};
 </script>
@@ -37,5 +52,22 @@
 	name="With HTML content"
   args={{
     value: "Enim id qui <a href='https://urban.org' target='_blank'>labore labore quis</a> ut enim tempor sint quis proident voluptate ex. Duis nisi minim et occaecat do ullamco nisi dolore ipsum proident tempor aute exercitation duis. Proident pariatur consectetur tempor mollit Lorem deserunt. Ut laboris est cillum ut dolore quis consectetur nostrud ut cupidatat enim ea cupidatat ipsum. Voluptate esse labore incididunt sint eiusmod ullamco mollit consequat.",
+  }}
+/>
+
+<Story
+	name="Reversed variant"
+  parameters={{backgrounds: {default: "dark"}}}
+  args={{
+    value: "Enim id qui <a href='https://urban.org' target='_blank'>labore labore quis</a> ut enim tempor sint quis proident voluptate ex. Duis nisi minim et occaecat do ullamco nisi dolore ipsum proident tempor aute exercitation duis. Proident pariatur consectetur tempor mollit Lorem deserunt. Ut laboris est cillum ut dolore quis consectetur nostrud ut cupidatat enim ea cupidatat ipsum. Voluptate esse labore incididunt sint eiusmod ullamco mollit consequat.",
+    variant: "reverse"
+  }}
+/>
+
+<Story
+	name="With color override"
+  args={{
+    value: "Enim id qui <a href='https://urban.org' target='_blank'>labore labore quis</a> ut enim tempor sint quis proident voluptate ex. Duis nisi minim et occaecat do ullamco nisi dolore ipsum proident tempor aute exercitation duis. Proident pariatur consectetur tempor mollit Lorem deserunt. Ut laboris est cillum ut dolore quis consectetur nostrud ut cupidatat enim ea cupidatat ipsum. Voluptate esse labore incididunt sint eiusmod ullamco mollit consequat.",
+    color: "#0a4c6a"
   }}
 />